+++ model/withdrawal.go
@@ -7,7 +7,7 @@ import (
 
 type Withdrawal struct {
 	Order       string    `json:"order"`
-	Sum         float32   `json:"sum"`
+	Sum         float64   `json:"sum"`
 	ProcessedAt time.Time `json:"-"`
 }
